Storage/Download_python/requirements.txt
google-cloud-error-reporting==0.30.0
google-cloud-storage
